minor refactor on api route
